requirements.txt
requests==2.25.1
Flask==2.0.0
mysql_connector_python==8.0.25
python_decouple==3.4.0